The behavioral update
Now everything can have subbehaviors.
How it works is that the weight of a node is calculated by having first it's own multiplier *  the calculated multiplier from the kind of node it is * the total sum of weights from each sub Behavior node.
That should give us much more  leeway in changing a few of its behaviors.