Brave new world
- The AI Behavior now takes into account the multiple targets it's hitting
- The archer is now "Complete", problems with the AI would be sometimes hitting own team mates with the Quick Shoot due to allies not being filtered out and due to the logic pairing them up with the enemy without duplicating the enemy as a target due to low health. Basically if the enemy is low on health, big chance the archer will hit an ally if the archer is a bit far away.
- Added behavior nodes based on own attributes or target attribute. The way it works that it has 4 key values.
Att To and from are the range this node is going to compare against.
Wto and from are the weights assigned to each end of the range.
What it'll do is match what the value currently is to the weight.
default weights are 0 to 1. the archer attfrom is 50 to 0, which means if the target has 50 health it weighs it at 0, then if target has 0 health, it weighs it at 1,
this essentially means that the archer will rate targets in range with lower health higher than others.
- UNITS FINALLY CAN JOIN
- Updated AI logging info to better reflect what's happening